Make listings ignore margins by default
Normally the listings environment will take margins, such as indents due to
enumerations, into account when adding leading space to the beginning of
formatted lines. However, this fails when the output is placed in a TikZ node
as the indent is set relative to the left side of the node instead of the left
side of the paper.

So, `ignoremargins=true` is now hardcoded. Need to find a way of influencing
the indent of the frames as they are currently centered on the page if
`\linewidth` happens to be less than `\textwidth`.