Use \lstnewenvironment instead of custom verbatim
Previously the `tikzCodeBlock` environment initiated a custom verbatim
environment that piped all output to a temporary file and read it back in using
`\lstinputlisting`.  This was because `\begin{listings}` does not work when
embedded inside another environment and the package author was ignorant of how
to properly use `\lstnewenvironment`.

`tikzCodeBlock` is now defined using `\lstinputlisting` which brings two
benefits:

  - There is no longer a requirement on the `verbatim` package and no need for
    a bunch of custom code that implements verbatim capture of input.
    Importantly, there is no need to allocate a temporary file to hold the
    input.

  - `tikzCodeBlock` is now better behaved with respect to optional arguments.
    Previously an empty options bracket, `[]`, always had to be added or the
    environment would gobble something from the beginning of its contents.